storefront2/requirements.txt
Django>=4.2.5
django-debug-toolbar>=4.0
pillow
mysqlclient
djangorestframework